requirements.txt
requests
python-telegram-bot
bs4
numpy
